Reverted back to using a offset in the database for the new strand, since keeping existing data is way easier this way and prevents having to drop/migrate tables

+++ src/main/java/de/tim_greller/susserver/persistence/repository/GameProgressionRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 public interface GameProgressionRepository extends
-        JpaRepository<GameProgressionEntity, Long>,
+        JpaRepository<GameProgressionEntity, Integer>,
         JpaSpecificationExecutor<GameProgressionEntity> {
 
     @Query("SELECT gp FROM GameProgressionEntity gp WHERE gp.mode = :mode ORDER BY gp.orderIndex LIMIT 1")
